requirements.txt
bosdyn-client 
bosdyn-mission 
bosdyn-choreography-client 
bosdyn-orbit
PySide2
PySide6
opencv-python
open3d
matplotlib
pandas
scipy
numpy
openpyxl
scikit-learn
qimage2ndarray
opcua
cryptography
psutil
pyqtgraph
pylibdmtx
pyzbar
pyopengl